Add guards to fail points for more graceful error recovery - ID: 3428871
Changes don't throw more exception, instead they add file exists guards
to prevent the need to throw a FileNotFoundException. That is, process
will fail early before unrecoverable exceptions will occur.

+++ src/org/dita/dost/writer/TopicRefWriter.java
@@ -550,6 +550,12 @@ public final class TopicRefWriter extends AbstractXMLWriter {
                 file = filename;
             }
             inputFile = new File(file);
+            if (!inputFile.exists()) {
+                final Properties prop = new Properties();
+                prop.put("%1", file);
+                logger.logError(MessageUtils.getMessage("DOTX008E", prop).toString());
+                return;
+            }
             outputFile = new File(file + FILE_EXTENSION_TEMP);
             fileOutput = new FileOutputStream(outputFile);
             ditaFileOutput = new OutputStreamWriter(fileOutput, UTF8);
@@ -575,10 +581,12 @@ public final class TopicRefWriter extends AbstractXMLWriter {
         } catch (final Exception e) {
             logger.logException(e);
         } finally {
-            try {
-                fileOutput.close();
-            } catch (final Exception e) {
-                logger.logException(e);
+            if (fileOutput != null) {
+                try {
+                    fileOutput.close();
+                } catch (final Exception e) {
+                    logger.logException(e);
+                }
             }
         }
     }
